monarch_app/backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
pandas==1.5.2
numpy==1.24.1
matplotlib==3.6.2
sqlalchemy==2.0.5
psycopg2==2.9.3
python-dotenv==0.21.0
passlib==1.7.4
pydantic==2.0.2
bcrypt == 4.3.0
PyJWT==2.8.0
python-multipart==0.0.6
# google-generativeai==0.3.0